spark/gui/requirements.txt
streamlit
streamlit-folium
folium
sqlalchemy
psycopg2-binary
kafka-python
